+++ packages/models/src/chat/message/content.ts
@@ -16,7 +16,7 @@ export const CodeChangeBlockSchema = z.object({
 
 export const UserContentBlockSchema = TextBlockSchema;
 
-export const AssistantContentBlockSchema = z.union([TextBlockSchema, CodeChangeBlockSchema]);
+export const AssistantContentBlockSchema = z.discriminatedUnion('type', [TextBlockSchema, CodeChangeBlockSchema]);
 
 export type TextBlock = z.infer<typeof TextBlockSchema>;
 export type CodeChangeBlock = z.infer<typeof CodeChangeBlockSchema>;
